First cut of project system
Adds a proper project system for Razor. Since we depend on the Roslyn
workspace as well as the underlying project system, we need to wait for
both the 'host' project and the Roslyn project before we can consider a
project 'ready'.

Note that this is a prototype, and I don't attempt to handle
cross-targetting.