Classify non-C# inside C#
Fixes a bug with preview formatting for FAR.

So when we ask the Roslyn API to classify C# for us, it will only
classify the actual C# tokens. We are responsible for filling in the
gaps and whitespace.

The bug is that the following text would have all of its whitespace
removed in the VS FAR preview window.

```
@{ var foo = "Hello, world!"; }
```

Would look like:
```
@{varfoo="Hello, world!";}
```

This fixes the issue and makes it look like what one would expect.